add import load indicator
