+++ app/(dashboard)/alerts/components/NotificationsTable.tsx
@@ -8,7 +8,7 @@ import ServiceNameWithLogo from "src/components/ServiceNameWithLogo/ServiceNameW
 import { useGlobalData } from "src/providers/GlobalDataProvider";
 import { EventType } from "src/types/event";
 import { ResourceInstance } from "src/types/resourceInstance";
-import formatDateUTC from "src/utils/formatDateUTC";
+import formatDateLocal from "src/utils/formatDateLocal";
 
 import NotificationsTableHeader from "./NotificationsTableHeader";
 import MessageInput from "src/components/MessageInput/MessageInput";
@@ -49,7 +49,7 @@ const NotificationsTable = () => {
         const upgradePath = instance.maintenanceTasks?.upgrade_paths?.[0];
         const status = upgradePath?.upgrade_status;
         const scheduledTime = upgradePath?.upgrade_path_scheduled_at;
-        const message = `Upgrade is ${statusMap[status] || status}${scheduledTime ? " at " + formatDateUTC(scheduledTime) : ""}`;
+        const message = `Upgrade is ${statusMap[status] || status}${scheduledTime ? " at " + formatDateLocal(scheduledTime) : ""}`;
         return {
           ...instance,
           message,
