+++ app/(dashboard)/instances/components/StatusCell.tsx
@@ -3,7 +3,7 @@ import { styled, Tooltip, tooltipClasses } from "@mui/material";
 
 import { Text } from "src/components/Typography/Typography";
 import { UpgradeStatus } from "src/types/resourceInstance";
-import formatDateUTC from "src/utils/formatDateUTC";
+import formatDateLocal from "src/utils/formatDateLocal";
 
 type StatusCellProps = {
   upcomingUpgrade?: any;
@@ -113,7 +113,7 @@ const StatusCell: React.FC<StatusCellProps> = ({ upcomingUpgrade }) => {
           <div className="flex items-center gap-1">
             <CalendarIcon />
             <Text size="xsmall" weight="regular" color="#181D27" ellipsis>
-              {formatDateUTC(upcomingUpgrade.upgrade_path_scheduled_at)}
+              {formatDateLocal(upcomingUpgrade.upgrade_path_scheduled_at)}
             </Text>
           </div>
         ),
